+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.142.0] - Not released
+## [1.142.0] - 2025-11-30
 ### Added
 - Show found users and groups on the search results page.
 - Show previews of media links in comments (disabled by default). Supported link